Project1

+++ project 1/backend/src/controllers/authController.ts
@@ -168,7 +168,7 @@ export class AuthController {
       const resultUser = await pool.query(`SELECT * FROM users WHERE id = $1`, [userId]);
 
       if (resultUser.rows.length === 0) {
-        res.status(400).json({ message: 'We do not have a user for this id' });
+        res.status(400).json({ message: 'User not found' });
         return;
       }
 
@@ -267,7 +267,7 @@ export class AuthController {
       }
 
       const user = resultUser.rows[0];
-      await pool.query(`DELETE FROM users WHERE id = $1`);
+      await pool.query(`DELETE FROM users WHERE id = $1`,[userId]);
 
       res.status(200).json({
         message: 'User deleted successfully',
